Extracting current flow into separate states. I've placed it under `flow` in the hope that we can use this namespace to migrate the current flow state, e.g., modified flow steps, default flow, etc here.
