+++ client/lib/signup/progress-store.js
@@ -23,6 +23,7 @@ import {
 	submitStep,
 } from 'state/signup/progress/actions';
 import { getSignupProgress } from 'state/signup/progress/selectors';
+import { getCurrentFlow } from 'state/signup/flow/selectors';
 import SignupDependencyStore from './dependency-store';
 
 const SignupProgressStore = {
@@ -82,10 +83,6 @@ const SignupProgressStore = {
 	},
 };
 
-function addTimestamp( step ) {
-	return { ...step, lastUpdated: Date.now() };
-}
-
 function addStorableDependencies( step, action ) {
 	const unstorableDependencies = get( steps, [ step.stepName, 'unstorableDependencies' ] );
 
@@ -103,7 +100,7 @@ function addStorableDependencies( step, action ) {
  */
 SignupProgressStore.dispatchToken = Dispatcher.register( function( payload ) {
 	const { action } = payload;
-	const currentFlow = get( SignupProgressStore.reduxStore.getState(), 'signup.currentFlow', null );
+	const currentFlow = getCurrentFlow( SignupProgressStore.reduxStore.getState() );
 	const step = { ...action.data, lastUpdated: Date.now() };
 
 	Dispatcher.waitFor( [ SignupDependencyStore.dispatchToken ] );
@@ -117,18 +114,28 @@ SignupProgressStore.dispatchToken = Dispatcher.register( function( payload ) {
 	switch ( action.type ) {
 		case 'SAVE_SIGNUP_STEP':
 			SignupProgressStore.reduxStore.dispatch(
-				saveStep( addStorableDependencies( {
-					...step,
-					lastKnownFlow: currentFlow,
-				}, action ) )
+				saveStep(
+					addStorableDependencies(
+						{
+							...step,
+							lastKnownFlow: currentFlow,
+						},
+						action
+					)
+				)
 			);
 			break;
 		case 'SUBMIT_SIGNUP_STEP':
 			SignupProgressStore.reduxStore.dispatch(
-				submitStep( addStorableDependencies( {
-					...step,
-					lastKnownFlow: currentFlow,
-				}, action ) )
+				submitStep(
+					addStorableDependencies(
+						{
+							...step,
+							lastKnownFlow: currentFlow,
+						},
+						action
+					)
+				)
 			);
 			break;
 		case 'PROCESS_SIGNUP_STEP':
