+++ client/state/signup/flow/actions.js
@@ -1,0 +1,13 @@
+/** @format */
+
+/**
+ * Internal dependencies
+ */
+import { SIGNUP_CURRENT_FLOW_SET } from 'state/action-types';
+
+export function setCurrentFlow( flowName ) {
+	return {
+		type: SIGNUP_CURRENT_FLOW_SET,
+		flowName,
+	};
+}
