+++ client/state/signup/flow/selectors.js
@@ -1,0 +1,11 @@
+/** @format */
+
+/**
+ * External dependencies
+ */
+
+import { get } from 'lodash';
+
+export function getCurrentFlow( state ) {
+	return get( state, 'signup.flow.currentFlow', '' );
+}
